# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: Descubre República Dominicana — API
  version: 1.0.0
  description: API REST del portal turístico Descubre RD. Versionada en /api/v1. Autenticación via Bearer JWT (Supabase Auth).
  contact: 
    name: Equipo Técnico Descubre RD
    email: "tech@descubrerd.do"
  license: 
    name: Privativo — uso interno
servers:
  - url: "https://api.descubrerd.do/api/v1"
    description: Producción
  - url: "https://staging-api.descubrerd.do/api/v1"
    description: Staging
  - url: "http://localhost:3000/api/v1"
    description: Desarrollo local
...
components:
  securitySchemes: 
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT emitido por Supabase Auth. Header: Authorization: Bearer <token>"
  schemas: 
...
    FiscalInvoice: 
      type: object
      properties: 
        id: 
          type: string
          format: uuid
        invoice_number: 
          type: string
          example: INV-2026-A1B2C3D4
        ncf: 
          type: string
          example: B0200000001
        ncf_type: 
          type: string
          enum: 
            - B01
            - B02
            - B14
            - B15
            - E31
            - E32
            - E44
            - E45
        buyer_name: 
          type: string
        subtotal: 
          type: number
        itbis: 
          type: number
        total: 
          type: number
        currency: 
          type: string
          example: DOP
        security_code_dgii: 
          type: string
        qr_url: 
          type: string
          format: uri
        status: 
          type: string
          enum: 
            - issued
            - voided
            - annulled
        issued_at: 
          type: string
          format: date-time
    EventTicket: 
      type: object
      properties: 
        id: 
          type: string
        event_id: 
          type: string
          format: uuid
        user_id: 
          type: string
          format: uuid
        tier_name: 
          type: string
          example: VIP
        price_paid: 
          type: number
        currency: 
          type: string
        qr_code_hash: 
          type: string
          example: DR-TKT-ABCD1234EFGH5678
        status: 
          type: string
          enum: 
            - ISSUED
            - CHECKED_IN
            - CANCELLED
        checked_in_at: 
          type: string
          format: date-time
          nullable: true
paths:
...
  /creators/onboarding: 
    post: 
      tags: 
        - creadores
      summary: Registro de nuevo creador de contenido UGC
      security: 
        - bearerAuth: []
      responses: 
        201: 
          description: Perfil de creador registrado
  /creators/feed: 
    get: 
      tags: 
        - creadores
      summary: Feed publico de videos cortos de creadores
      parameters: 
        - name: category
          in: query
          schema: 
            type: string
        - name: destination
          in: query
          schema: 
            type: string
        - name: page
          in: query
          schema: 
            type: integer
      responses: 
        200: 
          description: Feed paginado de videos
  /creators/me: 
    get: 
      tags: 
        - creadores
      summary: Perfil y metricas del creador autenticado
      security: 
        - bearerAuth: []
      responses: 
        200: 
          description: Perfil del creador
  /creators/videos: 
    post: 
      tags: 
        - creadores
      summary: Publica un video UGC con slug unico
      security: 
        - bearerAuth: []
      responses: 
        201: 
          description: Video publicado
...
  /memberships/subscribe: 
    post: 
      tags: 
        - membresias
      summary: Suscribe al usuario a un plan de membresia
      security: 
        - bearerAuth: []
      responses: 
        201: 
          description: Membresia activada
  /memberships/me: 
    get: 
      tags: 
        - membresias
      summary: Membresia activa y balance de puntos del usuario
      security: 
        - bearerAuth: []
      responses: 
        200: 
          description: Estado de membresia
  /events/{id}/tickets/purchase: 
    post: 
      tags: 
        - ticketing
      summary: Compra entrada para un evento en vivo
      security: 
        - bearerAuth: []
      responses: 
        201: 
          description: Entrada emitida con QR
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/EventTicket"
  /events/tickets/verify: 
    post: 
      tags: 
        - ticketing
      summary: Verifica y hace check-in de entrada por codigo QR
      responses: 
        200: 
          description: Entrada verificada y check-in exitoso
        404: 
          description: QR no encontrado
        409: 
          description: Entrada ya utilizada
...
  /transport/book: 
    post: 
      tags: 
        - productos
      summary: "Reserva de traslado, chofer privado o rent-a-car"
      security: 
        - bearerAuth: []
      responses: 
        201: 
          description: Traslado reservado
...
  /invoices/issue: 
    post: 
      tags: 
        - facturacion
      summary: Emite comprobante fiscal electronico (e-CF DGII)
      security: 
        - bearerAuth: []
      responses: 
        201: 
          description: NCF emitido
          content: 
            application/json: 
              schema: 
                $ref: "#/components/schemas/FiscalInvoice"
  /invoices/{ncf}: 
    get: 
      tags: 
        - facturacion
      summary: Consulta comprobante fiscal por numero NCF
      parameters: 
        - name: ncf
          in: path
          required: true
          schema: 
            type: string
            example: B0200000001
      responses: 
        200: 
          description: Comprobante encontrado
        404: 
          description: NCF no existe
...